refactor(cli): isolate MCP task registry ownership
Refs #488
